Switch from requirements files to poetry
Previously `poetry lock` wasn't working due to having `python` declared
as a dependency in both the `[tool.poetry.dependencies]` and
`[tool.poetry.dev-dependencies]` sections of
[pyproject.toml](pyproject.toml]. @JamesLuaHMRC heroically figured out
that declaring in both is a pretty poor idead. So I'm just committing on
his behalf.

Further, since we're switching to poetry, this commit deletes
`requirements.txt` and `test-requirements.txt` files.

Also a `.tools-versions` file is being added so that asdf users (we
should all be using it!) are able to get the right tooling versions.